requirements.txt

aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-default-group==1.2.4
colorama==0.4.6
distro==1.9.0
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
ollama==0.4.7
openai==1.70.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
pydantic==2.11.2
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.2
six==1.17.0
sniffio==1.3.1
sqlite-fts4==1.0.3
sqlite-utils==3.38
starlette==0.46.1
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.0
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.34.0

sentence-transformers
faiss-cpu
